requirements.txt
httpx
ollama
python-dotenv
furhat-realtime-api